Do the line wrapping of Braille separately
Before it was done within the liblouis translation process. This works
ok for block elements but fairs poorly for inline elements. Now we use
a fairly crude hack: Add a space at the beginning of the line in the
xslt if wrapping is needed. A process which follows the xslt wraps the
lines that start with a space.