Fix the wrapping of long lines for sbsform
The input for sbsform needs to be formatted in such a way, that the
lines are not longer than 255 chars. For that reason we invoke the
textwrap.fill() function on the output of the xslt. Under some
circumstances this function splits the words at "non-word" characters.
Since the ASCII-Braille contains characters such as '*' or '-' which
are supposed to be part of the word, we can not use this assumption.
The parameter break_on_hyphens=False has the effect that splitting is
only done at white space.

Fixes #888
Fixes #873