Save the related data of words, i.e. documents and isConfirmed
When checking unknown words these are implicitely unconfirmed, so no
need to show them in the form. However they need to be marked as such
when saving. When confirming words they need to be marked as such and
the related document needs removing. Here we have an additional
complication that a form is not saved if it is not changed. So we show
the confirmed field in the form and (will) change it with some
Javascript.