Write Braille Tables when checking or confirming unknown words
The check, local and confirm requests now create braille tables. In
the case of check and local these aren't that big, so creation of the
tables doesn't take too long. There is probably also some room for
optimization when it comes to the SQL queries. However the confirm
request causes a regeneration of the global white list (for grade1 and
grade2), i.e. about 80M data is written to disk. This causes the
request to be slow. Again optimizing the SQL queries might help but
ultimately we'll have to think about some form of message queue to
handle this load asynchronously.