Added the production source info to the document
to store from which data source a document was derived. For now it can
be derived from an electronic source (e.g. a pdf) or be undefined. In
this context undefined probably means that it was scanned from a hard
copy. Other scenarios are possible in the future, e.g. OCR by a third
party or retrieved from Braille in the archive etc.

In detail the changes are as follows:

- Add a field ("production_source") to the document model
- Add a migration
- Allow prod:source in the name attribute of the meta element in the
  schema
- Add the meta tag to the initial XML 
- validate uploaded XML against the value of the field in the model
- Update the XML on changes in the model
- Grab the data source information from the ABACUS export when
  importing