Add some performance enhancements to the dictionary views.
Apparently a formset it still queries the db even if you pass data to
it. At least if not all of the forms in the formset are from the set
of initials. Now, if you do not pass a queryset it will query the
entire table (which is huge in our case). So you'll get big
performance boosts of you pass a smaller queryset.