Enable external hyphenation
This means that for Braille produced for sbsform is now pre-hyphenated
using libhyphen. For that some external dependencies have to be
installed (as documented in the INSTALL file). You also need a newer
version of the Liblouis Java bindings that can handle pre-hyphenated
text.

Also since we install yet another external tool the path to
dtbook2sbsform is now repurposed to just the path to external
utilities, which is used to find both dtbook2sbsform and
dtbook_hyphenator.