Make it possible to have the same unconfirmed word in multiple books
This is quite a far reaching change: The unique constraint is extended
to include the document. For that to be possible the ManyToMany
relationship between Word and Document was changed into a ForeignKey,
i.e. many words in a Document. This solves the immediate problem.
However it leaves the door open for duplicate entries in the global
dictionary. Since we set the document link to null when adding a word
to the global dictionary and since null != null according to the SQL
standard the unique index will not prevent two "identical" entries,
i.e. the same word with a null document entry. There are some
precautions to prevent this at the form level, however there are no
measures to prevent this on initial import of the global dictionary.

A better way to solve this would probably be to partition the table
horizontally, i.e. have separate tables for confirmed words, confirmed
local words and for unconfirmed words. This however would be more
effort to handle with regards to formsets, etc.

Also the UI is quite confusing, so maybe a better strategy would be to
just ignore the problem of duplicate entries and have a cron job which
cleans them out on a regular bases.