Split the word model into two models for global and other words
That way we can enable at the db level that no two entries for the
same word exist globally or even per book. There is a lengthy
discussion about this in http://redmine.sbszh.ch/issues/1683. In
essence the word table is split vertically into two tables one for
global, confirmed words. The uniqueness of these is guarantied using a
unique constraint at the database level. Words which are local to a
document (whether confirmed or not) are in a second table where we
also have a unique constraint to make sure the same word only exists
once per book. The downside is that we need to make two selects to
check for known words (when checking all words of a book). We need to
look at both tables where before it was one select from one table.