Fetch the product number from the abacus export
and use it when generating ebooks. This is done by appending the
product numbers to the form which lets the user import documents from
abacus. We're adding a second (invisible) formset that contains all
product-document (isbn really) mappings. This second formset is then
used when processing the POST request to update the product table.
When generating a ebook this table is consulted and a suitable ebook
number is suggested if found in the table.