Remove words from local dicts which aren't in the document
In some cases it can happen that a word is inserted in the local dict
and subsequently removed from the original document (maybe it was a
scan or ocr error). These words need to be removed from the local
words. This is now done on subsequent checks for unknown words.