Use iterator method on large querysets
According to
https://docs.djangoproject.com/en/dev/ref/models/querysets/#iterator
"a QuerySet typically caches its results internally so that repeated
evaluations do not result in additional queries". In other words if we
query all of the global words we have 1.2 million records and the
memory consumption goes up to 2 Gigs. The recommendation is as
follows: "For a QuerySet which returns a large number of objects that
you only need to access once, this can results in better performance
and a significant reduction in memory."