Make the isbn regexp in the abacus importer more compliant
Unfortunately there are two places where this needs to be changed.
There is the rnc which humans want to deal with and there is the rng
which is the only format accepted by lxml.etree. There is a Makefile
which automates the conversion but if you forget to invoke it you'll
end up with situations like this one.