Make the file uploads sequential
There seems to be a problem in that MySQL gets overwhelmed when
uploading images in parallel ('Deadlock found when trying to get lock;
try restarting transaction'). So the theory is that when throttling the
image uploads we can avoid this problem and sequential upload should
throttle the upload.

Of course we could implement multi file upload on the server side and
avoid the problem that way. But this approach is simpler and good enough
for our purposes.