Fix the image upload
Use standard methods to detect if an image has been uploaded previously,
i.e. do a get inside a try/catch and insert if the get failed otherwise
just update the file.

Also implement the storage properly so that the old file really is
overwritten.

This allows is to get rid of the weird save method that the Image class
had before where we'd delete any existing object before saving the new
one.