Delete images when the production is finished
Images take up a lot of space and are used mostly for productions that
are only used once. Since images are archived in other places and we
most likely will not be using them again after a production is finished,
we can safely remove them.

If a production is started again we will have to upload the images
again.