Use sbs-de-core.mod when translating with grade2
For some reason the code was removing the sbs-de-core.mod when
translating for grade 2. This would lead to different results than the
translation at run time (which uses sbs-de-core.mod). But since the
result was correct against the dictionary the word was not added to
the white list. But then at run time (using sbs-de-core.mod) the word
was not translated correctly and not in the white list.

This did not matter in older liblouis versions but apparently does
since liblouis/liblouis@1f3c08b.

Related to #12