Show the hyphenations in the check_words_g2 view
This seems like a crude hack. It pulls the hyphenated words from a
REST API (that in itself is quite OK) and then adds it to some generic
data structure. Then the WordFormSet uses a different form when we are
in grade 2 and when building the form.

When submitting the values are ignored