Fix a lovely UnicodeDecodeError
When invoking a shell command you need to

1. invoke the shell command with the proper LANG environment set
2. encode the command according to above setting

In other words you need to encode the strings in the command before
shipping it to the subprocess.

If you neglect to do that then Python (2) will implicitly encode the
command with the ascii encoding which will of course fail if you have
anything but ascii in your command.

See also https://nedbatchelder.com/text/unipain.html for a very
enlightening treatise of this painful subject.